Translate comments from Friulan Italian to English
I decided to submit this code to this context: https://underthrow.substack.com/p/the-constitution-of-consent-contest

Therefore to abide the rules it shall be written in english, this is why Im rewriting the comments from Trieste furlane to english dialect. 

Also I am committing a new file which is the constitution preable.

This contexts will only allow a 80-words preamble and it's quite not enought. Therefore a more decent version of it will be commited.

I really hope whoever reads this contract - one true real social contract made by legit social engagement, free, expliclt, consentend and voluntary - find this contract and code inspiring and can reasonably implement a trust chain and a really free social community as we did to ourselves.

Maybe this is one of the many possible ways to reach exterritorial cypher panarchy: by means of public contest (!! lol)

Hope other communities find it inspiring and more counter-misfits and cypher punks join the panarchist approach.

Maybe I will write further documentation because the front end code is not uniform and some of us use it straight from the blockchain explorer or individual hooks in our SeaBSD exterritorial HOA.